feat: moved dialogs, sidebar,usersettings remindernotifs to UI changed auth created styles folder maybe not in the comments since it is empty
