+++ src/client/src/UI/ViewAttendeesDialog.tsx
