+++ src/client/src/components/Home.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import Sidebar from './Sidebar';
-import EventDialog from './EventDialog';
-import CreateEventDialog from './CreateEventDialog';
-import ReminderNotification from './ReminderNotification';
+import Sidebar from '../UI/Sidebar';
+import EventDialog from '../UI/EventDialog';
+import CreateEventDialog from '../UI/CreateEventDialog';
+import ReminderNotification from '../UI/ReminderNotification';
 import { useHomeDashboard, useReminders } from '../hooks/hooks';
 import '../styles/home.css';
 import '../styles/login-page.css';
